+++ inst/run/prod.R
@@ -1,0 +1,6 @@
+pkgload::load_all(
+	reset = TRUE,
+	helpers = FALSE
+)
+
+run()