+++ NEWS.md
@@ -4,6 +4,8 @@
 - Add `tag2` convenience function to htmltools plugin
 - Add `run_dev` function that runs the build step before
 the launching the app.
+- Changed `build` message when nothing is to be built
+- Added `add_app_file` function
 
 # leprechaun 1.0.0.9000